🚚 api: Move certification-attestation route to certification/results context
