+++ api/tests/certification/results/acceptance/application/sticker.pdf
