+++ api/tests/certification/results/unit/application/certification-attestation-controller_test.js
@@ -1,5 +1,5 @@
-import { certificationAttestationController } from '../../../../../src/certification/course/application/certification-attestation-controller.js';
-import { usecases } from '../../../../../src/certification/course/domain/usecases/index.js';
+import { certificationAttestationController } from '../../../../../src/certification/results/application/certification-attestation-controller.js';
+import { usecases } from '../../../../../src/certification/results/domain/usecases/index.js';
 import { LANGUAGES_CODE } from '../../../../../src/shared/domain/services/language-service.js';
 import { domainBuilder, expect, hFake, sinon } from '../../../../test-helper.js';
 import { getI18n } from '../../../../tooling/i18n/i18n.js';
