+++ api/tests/certification/results/unit/application/certification-attestation-route_test.js
@@ -1,4 +1,4 @@
-import * as moduleUnderTest from '../../../../../src/certification/course/application/certification-attestation-route.js';
+import * as moduleUnderTest from '../../../../../src/certification/results/application/certification-attestation-route.js';
 import { securityPreHandlers } from '../../../../../src/shared/application/security-pre-handlers.js';
 import { expect, HttpTestServer, sinon } from '../../../../test-helper.js';
 describe('Unit | Route | certification-attestation-route', function () {
